requirements.txt
fastapi
uvicorn
python-docx 
qdrant-client
sentence-transformers
websockets
python-dotenv
python-multipart
